+++ CHANGELOG.md
@@ -77,6 +77,7 @@ Changelog
 - Daedalus and Cardano node update notification text messages change ([PR 392](https://github.com/input-output-hk/daedalus/pull/392))
 - Fixed brittle acceptance test steps (wallet latest transaction amount check) ([PR 393](https://github.com/input-output-hk/daedalus/pull/393))
 - Fixed Numeric component caret position bug on wallet send screen ([PR 394](https://github.com/input-output-hk/daedalus/pull/394))
+- Show correct error message for A -> A transaction error ([PR ])
 
 ### Chores